+++ src/pages/cash-register/invoicing/Invoicing.module.css
@@ -28,39 +28,6 @@
   flex-grow: 1;
 }
 
-.name_price_wrapper {
-  display: flex;
-  justify-content: space-between;
-  font-weight: bold;
-  font-size: 17px;
-}
-
-.qty_rate_wrapper {
-  font-size: 1rem;
-}
-
-.product {
-  padding: 6px 8px;
-  cursor: pointer;
-}
-
-#total {
-  font-size: 18px;
-  font-weight: bold;
-  border-top: 2px solid var(--text-color);
-  padding: 7px 20px 0;
-}
-
-#total_wrapper {
-  display: flex;
-  justify-content: flex-end;
-  padding: 0 8px 1rem;
-}
-
-.selected_product {
-  background: aliceblue;
-}
-
 #header_container {
   height: 3rem;
   background-color: #5590bf;
@@ -76,18 +43,4 @@
   width: 130px;
   height: 100%;
   margin-left: 0.5rem;
-}
-
-#message_container {
-  font-size: 5rem;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-#message_container > span {
-  font-size: 1.3rem;
-  color: #bdbdbd;
 }
